requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
SQLAlchemy==2.0.21
Alembic==1.12.0
python-magic==0.4.27
boto3==1.28.62
botocore==1.31.62
Pillow==10.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
PyMySQL==1.1.0
cryptography==41.0.4
